Asiamat

An adjacency matrix representing the ground truth DAG used to generate a synthetic dataset from Lauritzen and Spiegelhalter (1988) about lung diseases (tuberculosis, lung cancer or bronchitis) and visits to Asia.

Asiamat

Format

A binary matrix with 8 rows and 8 columns representing an adjacency matrix of a DAG with 8 nodes:

D (dyspnoea), binary 1/0 corresponding to "yes" and "no"

T (tuberculosis), binary 1/0 corresponding to "yes" and "no"

L (lung cancer), binary 1/0 corresponding to "yes" and "no"

B (bronchitis), binary 1/0 corresponding to "yes" and "no"

A (visit to Asia), binary 1/0 corresponding to "yes" and "no"

S (smoking), binary 1/0 corresponding to "yes" and "no"

X (chest X-ray), binary 1/0 corresponding to "yes" and "no"

E (tuberculosis versus lung cancer/bronchitis), binary 1/0 corresponding to "yes" and "no"

Source

https://www.bnlearn.com/bnrepository/

References

Lauritzen S, Spiegelhalter D (1988). `Local Computation with Probabilities on Graphical Structures and their Application to Expert Systems (with discussion)'. Journal of the Royal Statistical Society: Series B 50, 157-224.